# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Go Clean Architecture.",
        "title": "Go Clean Arch API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
        "/pdf/merge": {
            "post": {
                "description": "Merge the provided list of PDF files into a single file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdfs"
                ],
                "summary": "Merge PDF files",
                "parameters": [
                    {
                        "description": "List of file names to be merged",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.MergeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The merged PDF file",
                        "schema": {
                            "$ref": "#/definitions/domain.Pdf"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity error details",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/pdf/upload": {
            "post": {
                "description": "Upload one or more PDF files and save them to the server",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdfs"
                ],
                "summary": "Upload PDF files",
                "parameters": [
                    {
                        "type": "file",
                        "description": "PDF files to be uploaded",
                        "name": "pdfs",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "List of uploaded PDFs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Pdf"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request, error details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Pdf": {
            "type": "object",
            "required": [
                "file_name",
                "file_path",
                "file_size"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "rest.MergeRequest": {
            "type": "object",
            "properties": {
                "fileNames": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}